Use the itemlaoders library (#4516)
