+++ setup.cfg
@@ -118,6 +118,9 @@ ignore_errors = True
 [mypy-tests.test_loader]
 ignore_errors = True
 
+[mypy-tests.test_loader_deprecated]
+ignore_errors = True
+
 [mypy-tests.test_pipeline_crawl]
 ignore_errors = True
