+++ setup.py
@@ -71,6 +71,7 @@ setup(
         'Twisted>=17.9.0',
         'cryptography>=2.0',
         'cssselect>=0.9.1',
+        'itemloaders>=1.0.1',
         'lxml>=3.5.0',
         'parsel>=1.5.0',
         'PyDispatcher>=2.0.5',
