+++ tests/requirements-py3.txt
@@ -1,7 +1,6 @@
 # Tests requirements
 attrs
 dataclasses; python_version == '3.6'
-jmespath
 mitmproxy; python_version >= '3.6'
 mitmproxy<4.0.0; python_version < '3.6'
 # https://github.com/pytest-dev/pytest-twisted/issues/93
